+++ modules/client/src/client-common/constants.ts
@@ -1,0 +1,14 @@
+import { ProposalMetadata } from "./interfaces/plugin";
+
+export const UNSUPPORTED_PROPOSAL_METADATA_LINK: ProposalMetadata = {
+  title: "(unsupported metadata link)",
+  summary: "(the link to the metadata is not supported)",
+  description: "(the link to the metadata is not supported)",
+  resources: [],
+};
+export const UNAVAILABLE_PROPOSAL_METADATA: ProposalMetadata = {
+  title: "(unavailable metadata)",
+  summary: "(the proposal metadata is not available)",
+  description: "(the proposal metadata is not available)",
+  resources: [],
+};